\name{OMearaLab-package}
\alias{omearalab-package}
\docType{package}
\title{
OMeara Lab utility and in-development functions
}
\description{
This project hosts scripts being developed by the O'Meara lab ( http://www.brianomeara.info ), typically for use in comparative methods. This is primarily an incubator for things in a raw state (before becoming projects of their own) or one-off scripts.
}
\details{
Some functions here are pretty useful (consensus tree with branch lengths, using the iPlant Taxonomic Name Resolution Service, etc. but use at your own risk.
}
\author{
Brian O'Meara
}